+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [v0.4] - Unreleased
-### added
-# 
+### Added
+* Add support for docker registry (#15)
 
 ## [v0.3] - 20180822
 ### Added